Clean up utils file
