+++ crates/core/src/types/number.rs
@@ -1,6 +1,7 @@
 //! Adapted from <https://github.com/YarnSpinnerTool/YarnSpinner/blob/da39c7195107d8211f21c263e4084f773b84eaff/YarnSpinner/Types/NumberType.cs>
 
-use crate::prelude::types::{type_util::*, TypeProperties};
+use crate::function_wrappers::*;
+use crate::prelude::types::TypeProperties;
 use crate::prelude::*;
 use std::ops::*;
